HealthComponent Split - Building lib-tempo
