+++ src/lib-tempo/CMakeLists.txt
@@ -11,7 +11,8 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationClient.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ID.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/Health.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentHealth.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/RenderHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/LevelManager.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/LevelRenderer.hpp"
@@ -38,7 +39,8 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationClient.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ID.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/Health.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentHealth.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/RenderHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/LevelManager.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/LevelRenderer.cpp"
