+++ src/lib-tempo/src/entity/SystemHealth.cpp
@@ -1,0 +1,21 @@
+#include <tempo/entity/SystemHealth.hpp>
+
+namespace tempo {
+
+	void SystemHealth::CheckHealth() {
+
+		auto entities = getEntities();
+
+		for (auto& entity : entities) {
+			auto& h = entity.getComponent<ComponentHealth>();
+			int id = entity.getId();
+
+			// If the entity has 0 health then kill entity
+			if ((h.current_health <= 0)){
+				printf("\nEntity ID: %d has just been \"killed\". \n", id);
+				// TODO: Sort something out for here
+				// entity.kill();
+			}
+		}
+	}
+}
